+++ src/components/Sparkline/Sparkline.tsx
@@ -58,7 +58,7 @@ export function Sparkline({
   }, 10);
 
   useLayoutEffect(() => {
-    if (!entry) return;
+    if (entry == null) return;
 
     if (containerRef == null) {
       throw new Error('No SVG rendered');